crosswalkr: Rename and Encode Data Frames Using External Crosswalk Files

A pair of functions for renaming and encoding data frames using external crosswalk files. Especially useful when constructing master data sets from multiple smaller data sets that do not name or encode variables consistently across files. Based on -renamefrom- and -encodefrom- Stata commands.

See also

Author

Maintainer: Benjamin Skinner btskinner@coe.ufl.edu (ORCID)