State and territory crosswalk data set.

An example state and territory crosswalk. Includes information for all states plus the District of Columbia plus territories.

sttercrosswalk

Format

A data frame with 69 rows and 10 variables:

stfips

Two-digit FIPS codes

stabbr

Two-letter abbreviation

stname

Full name

cenreg

Census region number

cenregnm

Census region name

cendiv

Census division number

cendivnm

Census division name

is_state

Indicator for status as state

is_state_dc

Indicator for status as state or DC

status

1 := Under U.S. sovereignty; 2 := Minor Outlying Islands; 3 := Independent nation under Compact of Free Association with U.S.; 4 := Individual Minor Outlying Islands (within status 2)